Refactoring to simplify the property test
